requirements.txt
sqlalchemy
flask
redis
redis-om
pydantic
nltk
requests
pandas
python-dateutil
spacy
pytz
parsedatetime
python-Levenshtein

flask_sqlalchemy
flask_login
